Fix an error where Appbridge would be initialized without a host
Previously when you made some changes to the apps codebase Appbridge could reinitialize without the host being preset in the URL.  This meant that Appbridge was reinitialized with undefined as a host.  This is invalid, which would then creash the app.

The solution here is to cache the host in session storage.  This way, even if the host is removed from the URL, the app can still access it from session storage.

To be clear, this should only ever be an issue in development mode.  In development mode, the application can code, and individual moduels can be swapped out at run time.  This can lead to Appbridge be reinitialized, after the host has been removed from the URL.  This should not be possible in production, because the code is static and should never change.  As such, in production, this is overkill.  But it has very little cost, and improves Dev X, so seems a worthy tradeoff